GS23: Change folder structure + amend router-controller funcitonality
