+++ Marketplace/src/app.js
@@ -1,0 +1,13 @@
+const express = require("express");
+const path = require("path");
+const rutaMain = require('./routers/main');
+const rutaProducto = require('./routers/product');
+
+const app = express();
+const staticFolder = path.resolve(__dirname, "../public");
+app.use(express.static(staticFolder));
+
+app.listen(3000, () => console.log('Servidor funcionando en puerto 3000...'));
+
+app.use('/', rutaMain);
+app.use('/product', rutaProducto);
