+++ Marketplace/src/controllers/mainController.js
@@ -1,0 +1,18 @@
+let path = require('path')
+
+const controlador = {
+	carrito: (req, res) => {
+		res.sendFile(path.resolve(__dirname, '../views/carrito.html'))
+	},
+	index: (req, res) => {
+		res.sendFile(path.resolve(__dirname, '../views/index.html'))
+	},
+	login: (req, res) => {
+		res.sendFile(path.resolve(__dirname, '../views/user/login.html'))
+	},
+	registro: (req, res) => {
+		res.sendFile(path.resolve(__dirname, '../views/user/registro.html'))
+	},
+};
+
+module.exports = controlador;
