+++ Marketplace/src/controllers/productController.js
@@ -1,0 +1,10 @@
+let path = require('path')
+
+const controlador = {
+	detalle: (req, res) => {
+		// this method should check the received id
+		res.sendFile(path.resolve(__dirname, '../views/product/detalle-producto.html'))
+	},
+};
+
+module.exports = controlador;
