+++ Marketplace/src/routers/main.js
@@ -1,0 +1,11 @@
+const express = require('express');
+const mainController = require('../controllers/mainController')
+
+const router = express.Router();
+
+router.get('/', mainController.index);
+router.get('/carrito', mainController.carrito);
+router.get('/login', mainController.login);
+router.get('/registro', mainController.registro);
+
+module.exports = router;
