+++ Marketplace/src/routers/product.js
@@ -1,0 +1,8 @@
+const express = require('express');
+const productController = require('../controllers/productController')
+
+const router = express.Router();
+
+router.get('/:id', productController.detalle);
+
+module.exports = router;
