--- src/2-routers/ruta-main.js
@@ -1,8 +1,0 @@
-const express= require('express');
-const controladorMain = require('../1-controllers/controlador-main')
-
-const router = express.Router();
-
-router.get('/:id', controladorMain.index);
-
-module.exports = router;
